New Glimmer-compatible API for gesture recognizer coordination
The utility methods exposed by this addon for coordinating gesture recognizers (preferRecognizer/stopPreferringRecognizer) relied on walking a component's `parentView` chain, but Glimmer components don't have a parentView property. This commit adds a `gesture` service with register/unregister methods. When a component/recognizer is registered, previously registered components are informed that the new recognizers is to be preferred.